t/command-line-flags/banderas-de-linea-de-comando/ closes #22
