+++ public/argumentos-en-la-linea-de-comandos
@@ -149,7 +149,7 @@ de comandos con banderas usando el paquete flag</p>
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="command-line-flags">Command-Line Flags</a>.
+        Siguiente ejemplo: <a href="banderas-de-linea-de-comandos">Banderas de línea de comandos</a>.
       </p>
       
       <p class="footer">
