+++ public/index.html
@@ -153,7 +153,7 @@
       
         <li><a href="argumentos-en-la-linea-de-comandos">Argumentos en la línea de comandos</a></li>
       
-        <li><a href="command-line-flags">Command-Line Flags</a></li>
+        <li><a href="banderas-de-linea-de-comandos">Banderas de línea de comandos</a></li>
       
         <li><a href="variables-de-entorno">Variables de entorno</a></li>
